move connector builder components into the same shared components/connectorBuilder directory
