+++ airbyte-webapp/src/components/connectorBuilder/StreamTestingPanel/LogsDisplay.module.scss
