+++ airbyte-webapp/src/components/connectorBuilder/StreamTestingPanel/ResultDisplay.module.scss
