+++ airbyte-webapp/src/components/connectorBuilder/StreamTestingPanel/SliceSelector.module.scss
