+++ airbyte-webapp/src/components/connectorBuilder/StreamTestingPanel/SliceSelector.tsx
