+++ airbyte-webapp/src/components/connectorBuilder/StreamTestingPanel/StreamSelector.module.scss
